pyproject.toml
[project]
name = "distill"
version = "0.1.0"
description = "Personal content curation platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.3.0",
    "python-dotenv>=1.2.2",
    "praw>=7.8.0",
    "pyyaml>=6.0.2",
    "youtube-transcript-api>=1.2.4",
]
